requirements.txt
Django==5.0.1
gunicorn==23.0.0
python-decouple==3.8
dj-database-url==2.1.0
whitenoise==6.6.0
openai>=1.55.0
google-generativeai>=0.8.0
django-allauth==0.61.0
requests==2.31.0
python3-openid==3.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
Pillow>=10.0.0
httpx>=0.27.0
python-telegram-bot>=20.0